Fixes pós-uso real: parser de decimais, CASH no snapshot, UX do fechamento
- parseDecimal: parser único em frontend/src/lib aceitando pt-BR e US (9911,06 / 9911.26 / 1.234,56 / 9,911.26); substitui replace(/\./g,'').replace(',','.') que multiplicava por 100 quando o user digitava "9911.26" em valor total. 13 unit tests. Aplicado em SnapshotItemEditModal, ManualPriceModal, PendencyPanel, BulkExtractReviewModal, DistributionComposer, MovementComposer, AssetModal, Accounts.
- Labels canônicos: "Preço unitário" e "Valor total" como únicos rótulos em tabelas/modais. "Preço fim" → "Preço unitário"; "Valor (BRL/USD)" → "Valor total (BRL/USD)".
- Coluna "Valor total (USD)" na tabela de posições congeladas do snapshot.
- SnapshotItemEditModal: Enter salva (Cmd+Enter no textarea de notas).
- PendencyPanel: grupos por FI sem pendência aberta ficam colapsados mas acessíveis — permite subir proventos pra FI cuja positions já está resolvida (XP estava sumindo após upload de positions). Header é toggle com chevron; sort: abertas → resolvidas → sem instituição.
- CASH no snapshot: asset_has_position(pos, asset) sempre retorna True pra asset_class=CASH. Os 5 "Saldo em Conta" estavam sumindo do snapshot novo porque qty=0 + invested=0; agora entram em todo Sincronizar com pendência pro user digitar o saldo.
- Sincronizar ativos: botão com spinner + disabled enquanto roda; substitui 3× alert() por toast inline (verde/cinza/vermelho) com auto-clear em 6s.

+++ frontend/src/components/AssetModal.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useMemo, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { api, type AccountOut, type AssetClass, type AssetOut, type AssetRequest, type FinancialInstitutionOut, type FixedIncomeDetails, type FixedIncomeIndexer, type AssetMovementOut, type PhysicalDetails, type PositionOut } from '../lib/api'
+import { parseDecimal } from '../lib/parseDecimal'
 import { useEscapeKey } from '../lib/useEscapeKey'
 
 const CLASS_LABELS: Record<AssetClass, string> = {
@@ -137,7 +138,7 @@ export default function AssetModal({ initial, institutions, forcedWorkspaceId, w
         currency,
         ticker: tickerForbidden ? null : (ticker.trim() || null),
         cnpj: isFund ? (cnpj.trim() || null) : null,
-        current_price: currentPrice.trim() === '' ? null : parseFloat(currentPrice),
+        current_price: currentPrice.trim() === '' ? null : parseDecimal(currentPrice),
         notes: notes.trim() || null,
       }
       if (showWorkspacePicker && workspaceId) {
@@ -151,8 +152,8 @@ export default function AssetModal({ initial, institutions, forcedWorkspaceId, w
           issue_date: fiIssueDate || null,
           maturity_date: fiMaturityDate,
           indexer: fiIndexer,
-          rate: fiRate === '' ? null : parseFloat(fiRate),
-          face_value: fiFaceValue === '' ? null : parseFloat(fiFaceValue),
+          rate: fiRate === '' ? null : parseDecimal(fiRate),
+          face_value: fiFaceValue === '' ? null : parseDecimal(fiFaceValue),
         }
       } else if (showPhysical) {
         payload.details = {
@@ -160,7 +161,7 @@ export default function AssetModal({ initial, institutions, forcedWorkspaceId, w
           city: phCity.trim() || null,
           state: phState.trim() || null,
           country: phCountry.trim() || null,
-          area_m2: phAreaM2 === '' ? null : parseFloat(phAreaM2),
+          area_m2: phAreaM2 === '' ? null : parseDecimal(phAreaM2),
           registration_number: phRegistration.trim() || null,
           make: phMake.trim() || null,
           model: phModel.trim() || null,
